% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_date_columns.R
\name{add_date_columns}
\alias{add_date_columns}
\title{Adds extra date columns to a dataframe}
\usage{
add_date_columns(date_time_df)
}
\arguments{
\item{date_time_df}{a dataframe with a date column called "date_time"}
}
\value{
The input dataframe with new columns for the yday, week, month 
and year
}
\description{
Adds extra date columns to a dataframe
}
