% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_gdd_base.R
\name{apply_gdd_base}
\alias{apply_gdd_base}
\title{Applies the base GDD level to an average daily raster}
\usage{
apply_gdd_base(raster, gdd_base)
}
\description{
Applies the base GDD level to an average daily raster
}
