% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brick_datetime_rasters.R
\name{brick_datetime_rasters}
\alias{brick_datetime_rasters}
\title{Adds rasters of datetime fields to brick}
\usage{
brick_datetime_rasters(brick, template_raster, date.now)
}
\description{
Adds rasters of datetime fields to brick
}
