% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_constant_raster_values.R
\name{extract_constant_raster_values}
\alias{extract_constant_raster_values}
\title{Extracts constant raster values to daily temperatures table}
\usage{
extract_constant_raster_values(temperatures_df, rasters_list,
  stations = stations_sp)
}
\arguments{
\item{temperatures_df}{a dataframe of temperature data. Must contain 'stationid' field}

\item{rasters_list}{a list of constant rasters used as input for modelling}

\item{stations_sp}{a SpatialPointsDataFrame object of weather stations. names() must be "stationid"
with AN1, AN2, AN3 ... etc.}
}
\value{
input temperatures_df with added columns, one for each raster in rasters_list
}
\description{
Extracts constant raster values to daily temperatures table
}
