% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_temporal_raster_values.R
\name{extract_temporal_raster_values}
\alias{extract_temporal_raster_values}
\title{Extracts raster values from a set of rasters in the temporal_raster_df}
\usage{
extract_temporal_raster_values(temporal_rasters_df, temperatures_df,
  col_name = "temporal_raster_value", stations = stations_sp,
  verbose = TRUE)
}
\value{
temperatures table with temporal raster values
}
\description{
Extracts raster values from a set of rasters in the temporal_raster_df
}
