% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_temporal_raster_df.R
\name{make_temporal_raster_df}
\alias{make_temporal_raster_df}
\title{Make dataframe for temporal rasters path and datetime}
\usage{
make_temporal_raster_df(in_folder, start_date, end_date, date_chars,
  date_format, extension = "tif")
}
\arguments{
\item{in_folder}{The folder location of a set of rasters}

\item{start_date}{Earliest date to include in output}

\item{end_date}{Last date to include in output}

\item{date_chars}{The characters in the file names that define the date}

\item{date_format}{The date format of the date_chars}
}
\value{
Dataframe with two columns: path to temporally varying rasters, and the date
}
\description{
Make dataframe for temporal rasters path and datetime
}
