% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_output.R
\name{model_output}
\alias{model_output}
\title{Models data and produces image outputs}
\usage{
model_output(max_model, min_model, max_alt_model, min_alt_model, output_folder,
  output_start_date, output_end_date, output_format = "GTiff",
  output_daily_max, output_daily_min, output_daily_avg, constant_rasters_list,
  temporal_rasters_df, temporal_var_name)
}
\description{
Models data and produces image outputs
}
