% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_model_predict_rasters.R
\name{predict_mean_daily_rasters}
\alias{predict_mean_daily_rasters}
\title{Generate daily mean temperature raster with GAM}
\usage{
predict_mean_daily_rasters(start_date, end_date, formula, alt_formula,
  temperatures_df, raster_brick, temporal_rasters_df = 1, output_folder,
  output_ext, verbose = TRUE)
}
\description{
Generate daily mean temperature raster with GAM
}
