% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swns_stations_df}
\alias{swns_stations_df}
\title{Weather station data in Southwest Nova Scotia}
\format{An object of class \code{data.frame} with 188163 rows and 7 columns.}
\usage{
swns_stations_df
}
\description{
A dataset containing the weather station temperature recordings and coordinates
}
\keyword{datasets}
