% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swns_stations_df_200}
\alias{swns_stations_df_200}
\title{Weather station data in Southwest Nova Scotia from 200m resolution}
\format{An object of class \code{data.frame} with 188163 rows and 15 columns.}
\usage{
swns_stations_df_200
}
\description{
A dataset containing the weather station temperature recordings, coordinates, and 
values extracted from environmental rasters at 200m resolution.
}
\keyword{datasets}
