% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_daily_GAMs.R
\name{validate_daily_GAMs}
\alias{validate_daily_GAMs}
\title{Provides validation data and statistics on daily GAMs}
\usage{
validate_daily_GAMs(model_stations_df, val_stations_df, years, days, formula,
  alt_formula = "", verbose = FALSE)
}
\description{
Provides validation data and statistics on daily GAMs
}
