% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_monthly_GAMs.R
\name{validate_monthly_GAMs}
\alias{validate_monthly_GAMs}
\title{Provides validation data and statistics on monthly GAMs}
\usage{
validate_monthly_GAMs(model_stations_df, val_stations_df, years, months,
  formula, alt_formula, verbose = FALSE)
}
\description{
Provides validation data and statistics on monthly GAMs
}
