% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_weekly_GAMs.R
\name{validate_weekly_GAMs}
\alias{validate_weekly_GAMs}
\title{Provides validation data and statistics on weekly GAMs}
\usage{
validate_weekly_GAMs(model_stations_df, val_stations_df, years, weeks, formula,
  alt_formula, verbose = FALSE)
}
\description{
Provides validation data and statistics on weekly GAMs
}
