% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeGDDout.R
\name{writeGDDout}
\alias{writeGDDout}
\title{Writes the gdd output in the function generate_gdd_output}
\usage{
writeGDDout(gdd_out, output_time_slice, gdd_base, date_now, plot_gdd_raster,
  output_folder, output_format)
}
\description{
Writes the gdd output in the function generate_gdd_output
}
