% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_gdd_base.R
\name{apply_gdd_base}
\alias{apply_gdd_base}
\title{Applies the base GDD level to an average daily raster}
\usage{
apply_gdd_base(raster, gdd_base)
}
\arguments{
\item{raster}{A daily average raster}

\item{gdd_base}{The set base temperature for setting GDD}
}
\value{
The raster with the gdd_base subtracted, and all negatives set to 0
}
\description{
Applies the base GDD level to an average daily raster
}
