% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_output_gdd.R
\name{generate_gdd_output}
\alias{generate_gdd_output}
\title{Generates accumulated GDD output rasters}
\usage{
generate_gdd_output(daily_avg_raster_df, gdd_base, start_date, end_date,
  output_time_slice, output_folder, output_format = "GTiff")
}
\arguments{
\item{daily_avg_raster_df}{output from make_temporal_raster_df with folder 
of the daily temperature average rasters}

\item{output_time_slice}{"daily", "weekly", "monthly", or "yearly"}
}
\description{
Generates accumulated GDD output rasters
}
