% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_df_700}
\alias{model_df_700}
\title{Temperatures and variables at 700m resolution}
\format{An object of class \code{data.frame} with 144270 rows and 15 columns.}
\usage{
model_df_700
}
\description{
Temperatures and variables at 700m resolution
}
\keyword{datasets}
